Pass down timeout to application tests
Summary:
The timeout value provided over grpc isn't being passed down to Application Tests as it is for Logic Tests.

Right now this isn't used in any way during the execution. However, D26662456 will address this. When no value in D26662456 is provided, it will revert to a 1hr default

Reviewed By: c-ryan747

Differential Revision: D26638051

fbshipit-source-id: 87c8b759cdbb03776090bb9972c7ffc16c5ba36a