Add Timeout & Sample to Application Tests
Summary:
In the same way that we do this for logic tests, we can now do this for Application Tests! The move to having chained futures makes applying timeouts on test execution far easier than it was before. In this case we just have to pass down the timeout appropriately.

Before this there wasn't even any respecting of timeouts within the execution itself. This now means that we do have proper timeouts at the test level.

Since we're adding timeouts where they weren't respected before, a sane default must be picked (currently 1 hour). This is so that this change does not cause tests to start failing after landing. It's deliberately permissive as user-defined timeouts can be layered on top later on.

Reviewed By: jbardini

Differential Revision: D26662456

fbshipit-source-id: e355bbb4d3882e58b86642361eeaf504d16cc3ff