Use launchPath string for Agent Launch Configuration
Summary:
I want to bring task launching/configuration alongside each other for sims and macOS.

This "Agent Binary" is mostly superflous, it existed when we needed to serialize this information in JSON. That is no longer important or needed. It also poses a problem for `PATH` interpolation (i.e. you must know the full path of the binary, rather than "launchctl")

Reviewed By: c-ryan747

Differential Revision: D26686808

fbshipit-source-id: 6b163234bbf3b6d74cdcc1e65f24f0f6a4044e62