use absolute path for xctrace
Summary:
We used to call `xcrun xctrace`. It means that we will call `xcrun` directly which will in turn find the path for `xctrace` and call it for us. That way, we will not be able to inject code to xctrace.

We need to inject code into `xctrace` because `xctrace` cannot find custom simulators (same thing happened for `xcodebuild`).

Reviewed By: c-ryan747

Differential Revision: D27246956

fbshipit-source-id: 78711e5525c0977bc237dfc82b5d43301517c144