Fix log stream termination
Summary:
Currently the way we detect when a log stream is done is when a write in the response stream fails, signalling that the client has died.

This works fine if theres is still more data in the log stream to be written, however if there's not then this method will never finish, leaking both a log stream process and a idb companion thread.

Instead we can use `ServerContext.IsCancelled` to check for the client going away. `[FBFuture onQueue:... resolveWhen:...]` will repeatedly poll this every ~0.1s.

Differential Revision: D27462449

fbshipit-source-id: ff62732d404a499dfff62e81bbc903c41c3e5088