add shim to FBXCTraceConfiguration
Summary:
xctrace does not recognize simulators from a custom device set, so we need to inject a shim to xctrace to change its behavior.

In this diff, added a shim property to `FBXCTraceConfiguration` to allow us to inject it into xctrace when invoking xctrace. Also deleted `nullable` in method implementation for cleanliness.

Reviewed By: c-ryan747

Differential Revision: D27353157

fbshipit-source-id: 1534e030bb275001e89839ea037e2aafbfe45e6e