Support TLS encrypted client channel
Summary:
Since we have added a possibility to listen on TLS enabled socket in IDB companion, this change allows us to connect to the companion using idb cli.

In the future we can also add a flag to idb cli to work with TLS enabled companion. For now only programmatic API is exposed (via build method in grpc/client.py)

Reviewed By: c-ryan747

Differential Revision: D27822850

fbshipit-source-id: b95484e3118cdda36282f3094963e69c5ce1e3a5