Simplified Constructor in FBApplicationLaunchConfiguration
Summary:
Removes a ton of the alternative methods that overlap and have varying behaviours. Now all values must be provided. Additionally, there wasn't a single constructor that took all of the arguments which necessitated the `withDebugger` methods for example.

Now we can just construct the object once and it is fully immutable. Invariants are now performed upon usage (i.e debugger is mutually exclusive with relaunch only if needed). This also means that there's no need for failing builder methods. Just defer it to the launch where we have fallibility already.

Reviewed By: c-ryan747

Differential Revision: D26688023

fbshipit-source-id: 6a6873438ff3f5afce97de093c208382aeca289e