Remove FBProcessOutputConfiguration
Summary:
This was needed when we would serialize the output in fbsimctl. However, this behaviour is now un-needed. We can just jump directly to using the class that is derived from the configuration, instead of having it as a middleman. This also brings the launch much closer inline with that of `FBTaskConfiguration`

`FBProcessIO` does a far better job here. There's no need to have a complex constructor for dealing with different types. We can just construct the appropriate `FBProcessInput` or `FBProcessOutput` class and stick this in.

Reviewed By: zeyadsalloum

Differential Revision: D26721866

fbshipit-source-id: 38dca69ffafac4723b31bb5710bacb036d4c7c23