FBTaskConfiguration/FBAgentLaunchConfiguration -> FBProcessSpawnConfiguration
Summary:
We now have two almost identical classes, the only difference is the manner of spawning. For `FBTask` we just ignore this and nobody constructs `FBTaskConfiguration` directly in any case.

This will mean that we can start to specify the launching of a task the same across sims and devices, this is very useful for example in the world of xctest execution

Reviewed By: zeyadsalloum

Differential Revision: D26725400

fbshipit-source-id: 2aa563da1295762ff43b9ae5172fe3c49a4f0292