Fix crashing on CoreSimulator logging in Xcode 13
Summary:
I haven't yet figured out why the `va_args` handling is different in more recent versions of CoreSimulator but it does. This means that the process will segfault. This is the `CoreSimulator` that is installed to the OS, rather than anything bundled with Xcode. As a result, it won't map to a specific version that is `xcode-select`'d rather the most recent Xcode version that has ever been installed.

The earliest version I've found that is the culprit is `757`, which is bundled with Xcode 13 Beta, so let's bypass the log interception in this case until there's a proper fix

Reviewed By: jbardini

Differential Revision: D29160934

fbshipit-source-id: 67e87fb695046085eb13f783f2bbc12ed89d870d