Remove FBTestDaemonConnection
Summary:
In looking at some problems related to Xcode 12.5, I've discovered that `FBTestDaemonConnection` is essentially doing nothing.

All it is doing is calling `_IDE_initiateSessionWithIdentifier:forClient:atPath:protocolVersion:` and then finishing. This is currently also done in `FBTestBundleConnection`. As such, this is a complete waste of time without doing anything to be an essential part of running tests. Not only that but the implementation of how to perform this exchange is slightly different (it doesn't handle the "legacy protocol" so it would fail in that case anyway.

Instead we can just let `FBTestBundleConnection` do what it needs, it is essentially responsible for doing the right thing in any case.

Reviewed By: jbardini

Differential Revision: D29162425

fbshipit-source-id: bf2634177e98bf5d2c1b6adb56320b347bdefb70