Remove legacy testmanagerd protocol usage
Summary:
There was at one stage handling for iOS 8.4 in https://github.com/facebook/idb/issues/263 and D3333334 (https://github.com/facebook/idb/commit/a21413a86a8732a3187a5a368e53b87e2d9c7ee4)

However, the now removed `FBTestDaemonConnection` never had this fallback, this means that if that configuration was used it would have failed in `FBTestDaemonConnection` instead. As a result this compatibility may well (or not) work for `FBTestBundleConnection` but it never would have for `FBTestDaemonConnection`. This allows us to conclude that iOS 8.4 support was broken at some point in the past and this code is completely unused. Therefore, this fallback can be fully deleted.

Reviewed By: jbardini

Differential Revision: D29162644

fbshipit-source-id: 1fdcc6a1a5bf4f2a2a4dfdfe4cbaf070ac477318