Revert D26725053: Check Exit Codes for "completed" future in FBTask
Summary:
D26725053 (https://github.com/facebook/idb/commit/1f43868b9b5e4dcab1aed28684dba972f8f6cba2) was an attempt to remove functionality from `FBTask` so that it could be delegated up and brung closer in line with simulator processes. The net result was that for consumers of `FBTask` that did not use `runUntilCompletion`, for example `xcodebuild` operations, the checking was removed.

This adds the exit code checking back into `FBTask` directly, so that `-[FBTask completed]` will resolve with exit code checking.

In the future I want to make this completely explicit so that `-[FBTask exitCode]` will always return the exit code. Another future can be used instead to perform checking automatically.

Differential Revision: D29302991

fbshipit-source-id: f991e45616a277dc85eb331c5167dc3988383daa