TEST_SHIMS_DIRECTORY overrides the Shims search path
Summary:
`TEST_SHIMS_DIRECTORY` is used to test local changes to the shimulators.

`findShimDirectoryOnQueue` requires both libShimulator and libMaculator to be on the same directory. If both Shims are not in `TEST_SHIMS_DIRECTORY`, it'll silently move on to the default Shim directories and will eventually find a shimulator without the local changes.

These changes make the xctest command fail if both shims are not present in `TEST_SHIMS_DIRECTORY`

Reviewed By: lawrencelomax

Differential Revision: D29431649

fbshipit-source-id: df7776ec85e19c7b26ee5f6b6d97b18e9384bc2e