FBLaunchedProcess exposes configuration from which it was launched
Summary:
Since this property is duplicated between `FBTask` and `FBSimulatorLaunchedProcess` it can be moved to the protocol itself. This also exposes that we shouldn't have had a `copy` on the configuration object for `FBSimulatorLaunchedProcess` (which itself no longer implements `copy`).

There's a further reduction of the number of properties needed in the implementation since `FBProcessIO` is contained within the configuration object so there's no need to pass it in multiple places, it can be referenced from the configuration object

Reviewed By: jbardini

Differential Revision: D29230038

fbshipit-source-id: 26a292637b40052eff59edc986b2ca381f5cf137