Avoid data races in translation object singleton
Summary:
There's currently a condition where due to the use of a singleton `AXPTranslator`, we end up fighting with each request setting a different "bridgeDelegateToken" across requests. This will happen when there are concurrent requests.

Since `AXPTranslator` is tokenised, we can use this to ensure that for all translation requests that we make to the system we key this back to a simulator. This object is itself wrapping the `AXPTranslator` singleton so will ensure that each wrapped request will pass the token down wherever it is needed. When requests are run concurrently with each other, they will now have their own token. The `bridgeDelegate` remains stable over time now.

Reviewed By: jbardini

Differential Revision: D29428277

fbshipit-source-id: 02fd3bdcda50070624d43a24b7f0ae5ff4e025a0