Split XCTestCommands
Summary:
There's currently a tiering of test level support for platforms that support shimming. In theory this should eventually be possible for devices, but for now we need to tier this out at least.

This also will help with later changes where we can remove some of the hoops that you have to jump through to perform a test execution

Reviewed By: jamescoggan

Differential Revision: D29452386

fbshipit-source-id: 11e5be40803f51912b801921fcc7b08f63ffd4a7