Remove dead SimulatorKit Encoder
Summary:
`SimulatorKit` hasn't had this encoder since Xcode 9. This is due to the move away from the very `SimStreamProcessorService` and instead having video recording work in-process within `simctl`.

The `simctl` implementation is pretty reliable for some time now, so it makes more sense to use that. This can always be revisited later. For the time being, it looks like `simctl` is using pixel buffers/ AVAssetWriter to generate a video in-process similar to how `FBSimulatorControl` worked before there was official support for this within Xcode itself.

Reviewed By: jamescoggan

Differential Revision: D29549696

fbshipit-source-id: 36882b26ab2c4d4dd4480698f47b9ed0d2ec86c3