Add configuration to FBXCTestReporter
Summary:
Within `idb` we have to have information about whether logs should be collected on idb's reporter.

However, this is not the same representation as the value that we get over the wire via grpc. Internally, the companion must create directories that act as a container of coverage, resultBundles and logDirectories since these are not provided over the wire.

In order to make this simpler and aid further changes. I'm extracting the configuration of this into it's own class. This prevents the need for multiple duplicated properties as these values are copied between classes.

Eventually I want to make the reporter created inside the where the run is created, but for now this at least helps ensure that all the values are copied around via this single class. This will also mean that I can make the configuration object readonly on the reporter

Reviewed By: jamescoggan

Differential Revision: D29680636

fbshipit-source-id: cfec9ad9f3f5150c5781a3ef751d1db890db08ad