Remove FBXCTestReporterWithFiles
Summary:
This protocol was only needed to inject state into a single reporter within the chain. In all other cases it is ignored or unused.

Instead of pushing this into a protocol that must be implemented everywhere, we can just pass it to the relevant reporter when the reporter is created. I want to make further improvements to how and when these are created, but for the time being this mimics the behaviour of the way that paths are injected into `FBIDBXCTestReporter`

Reviewed By: jamescoggan

Differential Revision: D29674208

fbshipit-source-id: 264350ed84f4316b6f3f313d17a88f1f61bfbcac