Add tail grpc call & companion implementation
Summary: Now that we have an interface for this it is implementable. Using a bidirectional stream, since stopping writing based on a client hangup is very error prone (see the implementation of the `log` command). This means that we can immediately detect that a client is gone and teardown afterwards.

Reviewed By: fgasperij

Differential Revision: D30015732

fbshipit-source-id: 41a70a98ec3679e224270ccdf1611b9902b80cd3