Remove -[FBSimulator containerApplication]
Summary:
There's a single place where the "Container Application" actually has a usage, that's in *focusing* a Simulator (bringing it to the foreground). I've not encountered any occurrences where the SimulatorApp UI is used for viewing Simulators outside of the default device set. The only benefit to having all of this complexity would be to focus amongst to concurrently running SimulatorApp instances, where one of them happens to be using a device set.

The reality is that if you have custom device sets, they are used for background automation without a UI. The usage of `focus` is just there to foreground the single SimulatorApp instance. We can now have better failure messages when there are two SimulatorApp instances running.

In a later diff, I'll remove the ability to launch SimulatorApp from FBSimulatorControl.

Reviewed By: jbardini

Differential Revision: D29986867

fbshipit-source-id: 5f587c0b192153477c99c9e799ed0a61dda1a025