Remove -[FBSimulator launchdProcess]
Summary:
There is no value in fetching this anymore. At one stage in the past this was about determining that a Simulator was booted, since there wasn't an API for verifying that a Simulator has been booted.

That has changed since Xcode 9, so there's really no need for this to exist. A ton of overhead can be removed

Reviewed By: jbardini

Differential Revision: D30095885

fbshipit-source-id: 29f3547081a6a5851e8ddb32f24a8a1e90e408b1