Add ProcessOutput so that it can be used elsewhere
Summary:
This factors out a `ProcessOutput` message that can then be used in a `spawn` call.

At the moment this means that we need to send both the old and the new way of representing a process's output in `LaunchResponse`. Once this has been rolled out fully, we can remove this handling within the companion

Reviewed By: jamescoggan

Differential Revision: D30097348

fbshipit-source-id: 8508c8bf1e83bc951bdfafd72311453cb299cc99