`idb focus` upon a default device set SimulatorApp will boot
Summary:
With the removal of SimulatorApp launching on boot, it now makes sense to have "focus" launch a SimulatorApp if there is not one launched.

The reason for this is that "focus" is there to bring the SimulatorApp to the foreground for the purpose of presenting to the user. This means that we can assume the intent is to display the SimulatorApp. In the case of a SimulatorApp with the default device set, this makes sense to just launch and activate SimulatorApp

Reviewed By: rgiese

Differential Revision: D30130798

fbshipit-source-id: 510cfbf012e8b766f0c169d3d3ab54e33e1cc1e6