Extract synchronous resolution of signal/exitCode futures
Summary:
One of the big benefits to to way that `FBTask` resolves all of it's public futures are that they are resolved at the same time as `statLoc`. Additionally, the future is stored rather than mapped, meaning that the values of completed futures from the perspective of the interface are available *immediately*.

Additionally, we can inject the IO detachment into the resolution of the stat_loc future, which further simplifies the flow.

Reviewed By: jamescoggan

Differential Revision: D30128077

fbshipit-source-id: 3c531c786e51e872deb7bd1aa8d5d0f7784b2e01