Resolve signal/exitCode Futures synchronously in FBSimulatorLaunchedProcess
Summary:
Now that we've extracted the synchronous nature of populating `exitCode`/`signal` from `FBTask`, we can use it within the context of `FBSimulatorLaunchedProcess`.

This means that the behaviour of unwrapping the bitmask is shared as well as resolving these futures immediately after they have been resolved.

This also allowed a massive simplification of the initialization/futures chain for where `FBSimulatorLaunchedProcess` is created. This is due to the changes in earlier diffs.

Reviewed By: jamescoggan

Differential Revision: D30130586

fbshipit-source-id: 086eb9c888349a09fcdcb66bf670f81de8ff7f75