Remove inconsistency of output in list-targets
Summary:
Due to the lowercasing of the state string in the describe GRPC call, a simulator can be both "Booted" and "booted". The former is the case in the output of `list-targets` when no companion is connected, the latter when a companion is connected.

The reason this occurs is in the merging logic within the implementation of list-targets, the output from the grpc describe call is preferred in place of the output of from the companion's tailer.

Reviewed By: jamescoggan

Differential Revision: D30246210

fbshipit-source-id: 075530c3808335dc9c8e12c35a9e24f68b915426