SEV Task S240582 : Emit consistently cased simulator `type` from `idb list-devices`
Summary:
A recent fix (D30246210 (https://github.com/facebook/idb/commit/77de8aa0a716abf4b2ad7ff02d51c5a5528bcce7)) for S239890 over-fixed the output of idb and also changed the casing for a different field altogether, the type field. As it happens, that particular field was previously always lowercase and tools do depend on that, including (but not necessarily limited to) the Mobile App Runner VSCode extension.

This rolls back the portion of the previous diff that was causing issues.

Reviewed By: scottrice

Differential Revision: D30257782

fbshipit-source-id: d891e7949a90aee981f19471fe3b911823a2827c