Move task execution to runUntilCompletionWithAcceptableExitCodes:
Summary:
This will help extract the exit code checking from an implicit behaviour of `-[FBTask completed]` and into concrete semantics.

The default value when no exit codes are defined in the builder is a set with a single exit code of 0 defined. This results in the following transformations:

- For builders where the exit code is defined, this is moved to `runUntilCompletionWithAcceptableExitCodes`
- For builders where no exit code is defined, the default of `[NSSet setWithObject:0]` is passed since it is he default
- For builders where exit code checking is suppressed, `nil` is passed.

All other builders that do not `runUntilCompletion` are left intact. These changes will be more complex as they will require moving the exit code checking to another component.

Reviewed By: jamescoggan

Differential Revision: D30191049

fbshipit-source-id: 8edf0d155372fb46ffbb5048249be209e89c32e9