Move FBSimulatorLaunchedProcess to FBLaunchedProcess
Summary:
The launching of simulator processes has been reduced down to effectively a value object container.

With that being the case, we can now extract `FBSimulatorLaunchedProcess` to a concrete implementation of `FBLaunchedProcess` and use this in place of it.

In a future diff we can have `FBTask` can inherit from this class

Reviewed By: jamescoggan

Differential Revision: D30277603

fbshipit-source-id: 9af01393ce416d992b1ed2e8912e76e46a57846d