Change FBTestReporterForwarder into FBTestReporterAdapter
Summary:
Remove confusing method forwarding and simplify `FBTestReporterForwarder` to only adapt calls to the reporter.

API Mediator is responsible for the interface with test bundle and handling. It seems reasonable to have the API mediator to handle calls from the test bundle and to decide what needs to be passed on to the test reporter.

The main motivation behind this change is to solve an issue where API Mediator was not receiving all calls from test bundle (through protocol `XCTestManager_IDEInterface`) because it was released when [this context](https://fburl.com/ud9mym9z) popped just after the bundle connection was established (and before the tests execution had finished). Reporting still worked because `FBTestReporterForwarder` was retained by the bundle connection.

Reviewed By: jamescoggan

Differential Revision: D30246908

fbshipit-source-id: 7eeaf250f45aa12ad4a47acf3147edc9e2994357