Use "Xcode" to launch UI target test app
Summary:
XCTest, on Xcode 12.5, has changed the internal implementation of `XCUIApplication` adding a new method to launch the UI test target application bypassing the IDE and requesting it directly to the "Platform" (iOS itself). This implementation is defined at `-[XCUIApplicationImpl _launchUsingPlatformWithArguments:environment:]`

The previous behaviour continues to be defined at `-[XCUIApplicationImpl _launchUsingXcodeWithArguments:environment:]` -- which is the method called when executing a UI test via Xcode.

Although the basic functionality of UI tests still works with `...UsingPlatform...`,  `_launchUsingXcodeWithArguments:environment` is the desired behaviour for idb: idb acts as the IDE/Xcode in the test execution (through the implementation of the protocols `XCTestManager_IDEInterface` or `XCTMessagingChannel_RunnerToIDE`) and receives a request to launch the test target app via `_XCT_launchProcessWithPath:bundleID:arguments:environmentVariables:`. idb can, then, setup the target app process as required to provide functionalities like logs and/or code coverage collection.

The decision between Xcode over Platform is ultimately based on having `XCUIApplicationImpl.path != nil` and  `XCUIApplication.prefersPlatformLauncher == NO`. To have those properly set, we need to fill in `XCTtestConfiguration.testApplicationDependencies` with the target app bundleId and the absolute path to the app bundle installed in the simulator.

Reviewed By: lawrencelomax

Differential Revision: D30246897

fbshipit-source-id: 829dada42fd9a51e8d54304f740aba6f5953cee2