Remove dependence local state file
Summary:
The usage of a tailer with a local state file is in principle a good idea, but in practice has proven to be a headache.

The way that things currrently work is that when a list of local targets is requested, a companion process is spawned in the background. The output of the companion tailer is redirected to a file, where it continues to write forever. Upon further invocations, this file is read and then presented to the user. On the face of it, this sounds simple. In practice this is not the case:

- If the xcode version changes the tailer will need to be re-spawned, but this does not currently happen. `idb kill` must be used to nuke the local state and re-start the tailer
- If the companion binary is upgraded, the above is also true
- A lockfile must be used to synchronize access to the process's output, which could also be a bottlekneck.
- The file location must be writable to begin with.
- The tailer has to be managed on each invocation and in some cases could be abandoned forever, wasting system resources.
- The above also makes injecting a new companion version a challenge.

There are so many edges here Instead, we can just list with the companion whenever we require the list of targets. There's only one downside to doing this, which is the performance of `list-targets`. However, in my testing, there's about a ~30% regression in the performance due to the added overhead of spawning a companion on each `list-targets` invocation. I consider this to be a preferential outcome for the sake of edge case removal and consistency. Instead of further investing in bugfixes and patches for the tailer approach, this time is much better spent on improving the performance of the companion launch such that it reports targets much faster.

Reviewed By: fgasperij

Differential Revision: D30764892

fbshipit-source-id: 3c74bba3666553072291ce39744ee9a5ec1117b7