Remove FBiOSTargetQuery
Summary: The only use for this is to query an array of targets by UDID and return the filtered array. This was more useful in the days of fbsimctl, where a query was used to construct the argparse. Now that is done, we can instead have very direct methods for doing exactly the same.

Reviewed By: jamescoggan

Differential Revision: D31204569

fbshipit-source-id: b9609f308ca8d12f0482d406243f4169fd8c558a