Enable --wait-for-debugger flag for app test
Summary:
Adding support of --wait-for-debugger for app test.
The app test workflow is:
Launch host app --> connect to host app --> launch tests in app process

If we launch host app in suspended mode and we connect to the host app this will fail because the connection needs a running app. To avoid the test execution to fail before a debugger is attached and resume the process we need to check if the launched app process is running before keep the execution.

The debug workflow is:
Launch host app paused --> wait for debugger to attach --> wait for user to resume app --> connect to host app --> launch test

Reviewed By: jbardini

Differential Revision: D31114204

fbshipit-source-id: 4a54f087561e3a8bbb6c69a90c244ed97b48a38b