Fix Log Operation Cancellation
Summary:
The parent diff goes a long way to explain why cancellation is so weird. In essence the problem boils down to the implementation of `-[FBLaunchedProcess statLoc]`. If `-[FBLaunchedProcess statLoc]` is cancelled in any way then the following `sendSignal` implementation will not work

```
- (FBFuture<NSNumber *> *)sendSignal:(int)signo
{
  return [[FBFuture
    onQueue:self.queue resolve:^{
      // Do not kill if the process is already dead.
      if (self.statLoc.hasCompleted) {
        return self.statLoc;
      }
      kill(self.processIdentifier, signo);
      return self.statLoc;
    }]
    mapReplace:@(signo)];
}
```

If `statLoc` is cancelled `statLoc.hasCompleted` will be true, bypassing the call to `kill`.

This means that if a cancelled `-[FBLaunchedProcess statLoc]` future *can never be killed*. There's some additional fixes to do in time, but for the log operation case, I've used `-[FBLaunchedProcess exitedWithCodes:]` which will actually prevent propagation of the cancellation to statLoc. As a result we can then piggy back on this behaviour and add adequate cancellation responders for performing the signal.

Reviewed By: jbardini

Differential Revision: D31305295

fbshipit-source-id: fe6217e531772b2b78c4eacd14f78738ad94ce9f