Remove -[FBFuture shieldCancellation]
Summary:
`-[FBTask sheldCancellation]` is very confusing and isn't even used in the correct way.

There are only two usages within the entire codebase and they were originally used to ensure that upon a future being cancelled, that it wouldn't call the default behaviour of the base future which is to sigterm the process. This behaviour was very surprising and is now removed, so this shielding is un-needed.

What should be intentional here is that cancelling `-[FBTask statLoc]` should **never** work. We want to ensure that the process exit information is always retained. At the moment this is not the case, cancelling `statLoc` will prevent the exit information from appearing (setting the result of a future after it has been cancelled will not return the result). This should be changed in time, but for now I want to remove the false sense of security that calling this method has.

Instead, I'm going to remove this method and make sure that `-[FBLaunchedProcess exitedWithCodes:]` does not cancel the underlying task futures. This is **truly** shielding and this can be achieved by creating a future that is resolved from another, bypassing the cancellation mechanism.

Reviewed By: jbardini

Differential Revision: D31339018

fbshipit-source-id: 6ef1f8225f6cc3108b884bfa2779dc89b71d839c