Remove UDID -> TargetType detection
Summary:
The only purpose of this is to attempt to divine if there's a valid UDID provided.

Since we'll run the UDID against the sets to determine if a target is present, this really doesn't help us fail any faster and is ultimately very fragile. If the UDID format changes over time, we'll have to keep updating this code.

Instead, just use the UDID against the target sets and then we'll get better failure messages in any case

Reviewed By: jbardini

Differential Revision: D31538460

fbshipit-source-id: dfacef65c9ed430c808b6bd7f55a8c3eb5861557