Simplify directories in image mounting
Summary:
This diff allows for the traversal of "mountable" images. We normalize the paths to contain the appropriate images and hide where they are coming from. Directories are traversable and perform normalization on the way.

The mounted directory is empty if nothing is mounted, otherwise contains the local equivalents of the mounted paths.

This has required a some re-working of the internals of the disk image commands internally, but it does make them more general. The only excpetion to this is `ensureDeveloperDiskImageIsMounted` which is effectively a helper of the common sane default.

Reviewed By: fgasperij

Differential Revision: D31165737

fbshipit-source-id: 86ec6b8b894933d0a4e517140f57f89310752bab