Fix image mounting error messaging
Summary:
I've found that mounting the same image twice is a no-op, and results in an exit 0. The failure case that we care about here is when the wrong iOS version is mounted so error accordingly.

This means that we can get rid of the "failIfMounted" argument since this should never be needed.

Reviewed By: jamescoggan

Differential Revision: D31682636

fbshipit-source-id: 1ac39d97a199531fd12f2e92038b51cd52cdef21