Save pid of implicitly spawned companion inside CompanionInfo
Summary:
In order to adequately track the pid location of spawned companions, these can be saved within the CompanionInfo itself. Since Companion Servers are the only processes that can be spawned by the idb cli, this will aid in ensuring that they are adequately cleaned up.

Since upgrading the CLI can result in the "pid" field not being present in the old version, we need to make sure to handle the fact that the `pid` field may not be present during the update. Since the parsing is only based on field presence, mixing and matching versions will not be an issue.

Reviewed By: jbardini

Differential Revision: D31954120

fbshipit-source-id: ba1b37a8aea8901c8a0b1ffe2d5516bf15692680