Kill spawned companions on idb kill
Summary:
This has been broken for some time, but now that we're saving pids alongside the spawned companions, we can ensure that the kill command now actually works.

This is now very simple to do, as when clearing the companions, we can return the list of companions that have been cleared. Each of these can then be inspected to confirm that they have pids or not and subsequently killed

Reviewed By: jbardini

Differential Revision: D31954214

fbshipit-source-id: 44e59b429127ff1702c3ba49a69f39a1b14d1d09