Make the listing results show the language agnostic method names
Summary:
Swift methods that throw are translated to Obj-C selector names with the suffix `AndReturnError`
by the automatic bridging performed to interop between the two languages. This caused a mismatch
between the names we were using to refer to the test methods in the list_bundle results and in the
xctest_run results. This change uses `XCTestCase`'s `languageAgnosticMethodName` to always use the
version without the suffix.

Reviewed By: jbardini

Differential Revision: D32017944

fbshipit-source-id: 2e49c5024829274a59af5553ea4c6d77fff26986