Partial rollback: Persist non-debuggable apps in idb-applications
Summary:
This is a partial rollback of D31685386 (https://github.com/facebook/idb/commit/e5cb063873be2fe4df418dfa98fa24dc073a7cda) which makes app persists regardless of make-debuggable flag value.

The main reason is that not persisting apps into idb-applications breaks integration for expanding placeholders in xctestrun files. In particular expand `__IDB_APPSTORAGE__`

This is a temporary rollback to avoid rolling back the entire diff stack and it is relatively safe to do this change. A better alternative is to instead expand the variable into the app bundle path.

Reviewed By: jbardini

Differential Revision: D32182253

fbshipit-source-id: 09267f4118467c885446f268f9a53b14ea531b3e