idb will bind on unix domain sockets by default
Summary:
For some time now, we've had implicit companion creation create companion servers on ports. This is bad for a number of reasons:

- If the bind is on the public interface, then there's a plaintext interface to the device over the network. Not good.
- A Unix Domain Socket allows for arbitrary naming rather than port numbers which are limited and numeric, this is handy since it allows for a naming convention where the udid of the target is incorporated into it.
- No need for binding on port 0 the reading the port back out, this will mean that idb will play much nicer with other applications that may bind on a specific port. This will prevent idb from ever stealing this port.

The location that a companion binds should be considered effectively private, since addressing these targets is based on passing a UDID rather than extracting a companion location due to the side-effect of a previous UDID-based invocation.

Reviewed By: jbardini

Differential Revision: D31645378

fbshipit-source-id: 60299f22cc7ee636c9770468a27bb26249a20772