Use FBAMDServiceConnection based stream reads/writes from debugserver
Summary:
This is needed for the debugserver when used against a disk image from > Xcode 12 to work. The existing "twisted pair" was using plain reads/writes, where the new debugserver is TLS'd. This means that we need to use the underlying `FBAMDServiceConnection` calls, which boil down to the appropriate read/write calls within the `AMDevice` APIs, which "do the right thing".

Not a huge issue to fix, it's just using the new reader/writer paradigms within `FBAMDServiceConnection` and wiring it into to the existing setup

Reviewed By: jamescoggan

Differential Revision: D31921512

fbshipit-source-id: 44d343ef180ded20788f2ac837a9ab0a45ccde19