FBVideoStream support for FBDataConsumerSync and Async variants
Summary: This diff relaxes protocol requirements for video stream consumer. If previously we required the consumer to conform to FBDataConsumerSync protocol (due to passing raw passing pointers without copying underlying data), now we allow the consumer to not conform to this protocol. In which case all of the underlying data will be copied

Reviewed By: lawrencelomax

Differential Revision: D32489993

fbshipit-source-id: 20317fd10b008304a6e9bddcdf14a9d5a50e6f91