Use com.apple.os_trace_relay for PID Listing
Summary:
This has a number of advantages:

- This will work for devices that do not have a valid xcode install, since this is based on a pure lockdown service, rather than one provided by Xcode's developer disk image
- This will work with a locked phone, since mounting a disk image is not required.

`com.apple.os_trace_relay` has a pretty simple API for obtaining the process listing. I can only assume that the existence of this is based on the fact that this service is intended to provide the listing of process ids to drive the console pid selection on the host side. In any case, we can take full advantage of this for better pid listing on devices.

Reviewed By: jbardini

Differential Revision: D32385909

fbshipit-source-id: a4a72ae54d0453558e1e4933c6a00a836a19d1bc