Move -[FBXCTestCommands transportForTestManagerService] to Extended XCTest Commands
Summary:
The implementation for `FBDevice` here is erroneous since it just wraps the raw socket, which doesn't take into account any TLS'ing. As a result this leads to a misleading impression that the internal xctest execution implementation could use this.

Instead, we can push this down to the "extended" protocol where there's other non-device implemented approaches. If and when device support is added to `XCTestBootstrap`, this can be re-evaluated.

Reviewed By: jbardini

Differential Revision: D32673661

fbshipit-source-id: b0a61688f73c82c6c14d303bf5b6f017c97c8e6e