Hide internal details in FBAMDServiceConnection
Summary:
This is no longer used or needed (or should be) since all calls should be wrapped in the relevant read/write methods. Otherwise it's easy to accidentally read/write in plaintext over a TLS'd connection by consuming the socket directly.

The only exception here is `AFC`. However, this isn't a massive issue as we can add a method that understands how to perform the wrapping on the `FBAMDServiceConnection` instance

Reviewed By: jbardini

Differential Revision: D32664472

fbshipit-source-id: fb094f03d1a0c660c6a73d16b574aee5b9d4d00c