Fix bug logic test failing wait for debugger
Summary:
When running `idb xctest run logic --wait-for-debugger` command with booted simulator it fails to wait for debugger to attach With message:
```
{"bundleName": "", "className": "", "methodName": "", "logs": [], "duration": 0.0, "passed": false, "crashed": true, "status": "crashed", "failureInfo": {"message": "Failed to wait test process (pid 39673) to receive a SIGSTOP: 'No child processes'", "file": "", "line": 0}}
```
"No child process" message is because when the logic test is launched with booted simulator it's a subprocess of the simulator and not of the companion.
Fixed issue checking the status of of generic pid instead of assuming it's a subprocess.

Reviewed By: jbardini

Differential Revision: D32647988

fbshipit-source-id: 95bb61ace6d4d00afe7bbc8b7d44accc974d9abb